Add files via upload
These are all the past and current versions of the mineSweeper solver. Version 0.2 is a mostly working version with a singular known bug involving correct probablities for certain tiles through a specific chain of events, however it will still likely come up with a solution. Version 0.3 is unfinished, and currently will not produce any proper output. It is instead focused on rewriting certain portions of the code to clean it up, replacing the Coord() class in 0.2